requirement.txt
Django==1.7.4
argparse==1.2.1
django-filter==0.9.2
djangorestframework==3.0.4
wsgiref==0.1.2
